Adding loading folder and moving current loading methods
